refactor(llm): centralize runtime orchestration and simplify provider paths
Add llm runtime/reasoning helpers, move gateway implementations under adapters, and route the main request flow through APIClient instead of maintaining parallel runtime behavior in LLMClient. Remove the legacy provider_adapters fallback, preserve compatibility shims where needed, and update the contract/consolidation task docs to reflect the new boundary and deferred Link/LiteLLM follow-up.
